UI improvements

+++ app/src/main/java/agency/five/codebase/android/weatherinfoapp/data/repository/WeatherInfoRepositoryImpl.kt
@@ -33,7 +33,7 @@ class WeatherInfoRepositoryImpl(
                 weatherInfoResponse.toWeatherInfo(
                     location = location,
                     isFavorite = favoriteLocations.any { it.location == location },
-                    isHome = homePlace().any { it.location == location },
+                    isHome = homePlace().mapLatest {it}.first().any { it.location == location },
                     hourlyWeather = hourly,
                     dailyWeather = daily
                 )
@@ -43,14 +43,19 @@ class WeatherInfoRepositoryImpl(
     override fun favoritePlaces(): Flow<List<FavoriteLocation>> =
         favoriteLocationDao.favorites().map {
             it.map { dbFavoriteLocation ->
+                var isHome = false
+                homePlace().map { homeLocations ->
+                    isHome = homeLocations.first().isHome
+                    Log.e("FAV LOC", "${dbFavoriteLocation.location} - $isHome")
+                }
                 FavoriteLocation(
                     location = dbFavoriteLocation.location,
                     lon = dbFavoriteLocation.lon,
                     lat = dbFavoriteLocation.lat,
                     temperature = 0,
-                    iconId = "",
+                    iconId = dbFavoriteLocation.iconId,
                     isFavorite = true,
-                    isHome = false,
+                    isHome = isHome,
                 )
             }
         }.shareIn(
@@ -59,42 +64,45 @@ class WeatherInfoRepositoryImpl(
             replay = 1
         )
 
-    override fun homePlace(): List<FavoriteLocation> {
-        var list = listOf(FavoriteLocation("", 0.0f, 0.0f, 0, "", true, isHome = true))
-        homeLocationDao.home().map {
-                list = it.map { dbHomeLocation ->
+    override fun homePlace(): Flow<List<FavoriteLocation>> =
+        homeLocationDao.home().map { dbHome ->
+            dbHome.map { dbHomeLocation ->
                     FavoriteLocation(
                         location = dbHomeLocation.location,
                         lon = dbHomeLocation.lon.toFloat(),
                         lat = dbHomeLocation.lat.toFloat(),
                         temperature = 0,
                         iconId = "",
-                        isFavorite = true,
+                        isFavorite = false,
                         isHome = true,
                     )
                 }
-            }
-        return list
-    }
+            }.shareIn(
+            scope = CoroutineScope(ioDispatcher),
+            started = SharingStarted.WhileSubscribed(1000L),
+            replay = 1
+            )
 
     override suspend fun addHomeLocation(location: String, lon: Double, lat: Double) {
-        homeLocationDao.deleteHome()
-        Log.e("CLICK ON ADD HOME", "Location: $location")
-        homeLocationDao.insertHomeLocation(
-            DbHomeLocation(
-                location = location,
-                lon = lon,
-                lat = lat,
+        if(!homePlace().first().any { homePlace -> homePlace.location == location }) {
+            homeLocationDao.deleteHome()
+            homeLocationDao.insertHomeLocation(
+                DbHomeLocation(
+                    location = location,
+                    lon = lon,
+                    lat = lat
+                )
             )
-        )
+        }
     }
 
-    override suspend fun addLocationToFavorites(location: String, lat: Double, lon: Double) {
+    override suspend fun addLocationToFavorites(location: String, lat: Double, lon: Double, iconId: String) {
         favoriteLocationDao.insertLocation(
             DbFavoriteLocation(
                 location = location,
                 lon = lon.toFloat(),
-                lat = lat.toFloat()
+                lat = lat.toFloat(),
+                iconId = iconId,
             )
         )
     }
@@ -103,13 +111,11 @@ class WeatherInfoRepositoryImpl(
         favoriteLocationDao.delete(location)
     }
 
-    @OptIn(ExperimentalCoroutinesApi::class)
-    override suspend fun toggleFavorite(location: String, lat: Double, lon: Double) {
-        favoritePlaces().mapLatest {
-            Log.e("CLICK", "did we get to here!?")
-            if(it.any { favoriteLocation ->  favoriteLocation.location == location })
-                removeLocationFromFavorites(location)
-            else addLocationToFavorites(location, lat, lon)
-        }
+    override suspend fun toggleFavorite(location: String, lat: Double, lon: Double, iconId: String) {
+        val favPlaces = favoritePlaces().first()
+        if(favPlaces.any { favPlace -> favPlace.location == location })
+            removeLocationFromFavorites(location)
+        else
+            addLocationToFavorites(location, lat, lon, iconId)
     }
 }
